monitor/2zoracle: parse swap rate as float (#1898)
## Summary of Changes
- Update monitor 2zoracle watcher to parse swap rate from response as
float64 instead of int64
- Fixes https://github.com/malbeclabs/doublezero/issues/1897

## Testing Verification
- Updated test coverage to include float swap rate in response

+++ controlplane/monitor/internal/2z-oracle/client.go
@@ -10,12 +10,12 @@ import (
 )
 
 type SwapRateResponse struct {
-	SwapRate     int64  `json:"swapRate"`
-	Timestamp    int64  `json:"timestamp"`
-	Signature    string `json:"signature"`
-	SOLPriceUSD  string `json:"solPriceUsd"`
-	TwoZPriceUSD string `json:"twozPriceUsd"`
-	CacheHit     bool   `json:"cacheHit"`
+	SwapRate     float64 `json:"swapRate"`
+	Timestamp    int64   `json:"timestamp"`
+	Signature    string  `json:"signature"`
+	SOLPriceUSD  string  `json:"solPriceUsd"`
+	TwoZPriceUSD string  `json:"twozPriceUsd"`
+	CacheHit     bool    `json:"cacheHit"`
 }
 
 type HealthResponse struct {
